requirements-test.txt
Django==1.7
djangorestframework>=2.4.3
pytest-django==2.6
pytest==2.5.2
pytest-cov==1.6
flake8==2.2.2
